chore: move demandSingleValue and demandOneOfOptions to core
