+++ packages/wrangler/src/core/index.ts
@@ -1,1 +1,2 @@
 export { defineAlias, defineCommand, defineNamespace } from "./define-command";
+export { demandOneOfOption, demandSingleValue } from "./helpers";
