+++ packages/wrangler/src/index.ts
@@ -38,12 +38,17 @@ import {
 import { devHandler, devOptions } from "./dev";
 import { workerNamespaceCommands } from "./dispatch-namespace";
 import { docsHandler, docsOptions } from "./docs";
-import { JsonFriendlyFatalError, UserError } from "./errors";
+import {
+	CommandLineArgsError,
+	JsonFriendlyFatalError,
+	UserError,
+} from "./errors";
 import { generateHandler, generateOptions } from "./generate";
 import { hyperdrive } from "./hyperdrive/index";
 import { initHandler, initOptions } from "./init";
 import "./kv";
 import "./workflows";
+import { demandSingleValue } from "./core";
 import { logBuildFailure, logger, LOGGER_LEVELS } from "./logger";
 import * as metrics from "./metrics";
 import { mTlsCertificateCommands } from "./mtls-certificate/cli";
@@ -158,55 +163,6 @@ export function getLegacyScriptName(
 		? `${args.name}-${args.env}`
 		: args.name ?? config.name;
 }
-
-/**
- * A helper to demand one of a set of options
- * via https://github.com/yargs/yargs/issues/1093#issuecomment-491299261
- */
-export function demandOneOfOption(...options: string[]) {
-	return function (argv: { [key: string]: unknown }) {
-		const count = options.filter((option) => argv[option]).length;
-		const lastOption = options.pop();
-
-		if (count === 0) {
-			throw new CommandLineArgsError(
-				`Exactly one of the arguments ${options.join(
-					", "
-				)} and ${lastOption} is required`
-			);
-		} else if (count > 1) {
-			throw new CommandLineArgsError(
-				`Arguments ${options.join(
-					", "
-				)} and ${lastOption} are mutually exclusive`
-			);
-		}
-
-		return true;
-	};
-}
-
-/**
- * A helper to ensure that an argument only receives a single value.
- *
- * This is a workaround for a limitation in yargs where non-array arguments can still receive multiple values
- * even though the inferred type is not an array.
- *
- * @see https://github.com/yargs/yargs/issues/1318
- */
-export function demandSingleValue(key: string) {
-	return function (argv: { [key: string]: unknown }) {
-		if (Array.isArray(argv[key])) {
-			throw new CommandLineArgsError(
-				`The argument '--${key}' expects a single value, but received multiple: ${JSON.stringify(argv[key])}.`
-			);
-		}
-
-		return true;
-	};
-}
-
-export class CommandLineArgsError extends UserError {}
 
 export function createCLIParser(argv: string[]) {
 	const experimentalGradualRollouts =
