+++ packages/wrangler/src/pubsub/pubsub-commands.ts
@@ -1,6 +1,6 @@
 import { readConfig } from "../config";
 import { confirm } from "../dialogs";
-import { CommandLineArgsError } from "../index";
+import { CommandLineArgsError } from "../errors";
 import { logger } from "../logger";
 import * as metrics from "../metrics";
 import { parseHumanDuration } from "../parse";
